Alteração da divisão

+++ src/com/dio/calculadora/Calculadora.java
@@ -17,7 +17,7 @@ public class Calculadora {
 		int soma = soma(a,b);
 		int subtracao = subtracao(a,b);
 		int multiplicacao = multiplicacao(a,b);
-		int divisao = divisao(a,b);
+		float divisao = divisao(a,b);
 		
 		System.out.println("Soma " + soma);
 		System.out.println("Subtração " + subtracao);
@@ -34,12 +34,11 @@ public class Calculadora {
 	}
 	
 
-	
 	public static int multiplicacao (int a, int b) {
 		return a * b;
 	}
 	
-	public static int divisao (int a, int b) {
+	public static float divisao (float a, float b) {
 		return a / b;
 	}
